Added Reviews Controllers to client/server and models. Added utils to handle Dtos and models mappping. TODO: Need to fix cookies not working and then Reviws endpoints can be tested. Cannot get CustomerId on my client because User is no initialized. TODO: Use Dtos in Services. TODO: Add explicit implicit opertaors or mapping from model-to-dto ones to every Dto.  TODO: Fix cookies auth when user logins + user logout does not work properly. TODO: cart does not work due to cookie not being created for cart.

+++ EcommerceRestApi/Models/Product.cs
@@ -16,9 +16,9 @@ public partial class Product : EntityBase
     [StringLength(30)]
     public string Code { get; set; } = null!;
 
-    public int Price { get; set; }
+    public decimal Price { get; set; }
 
-    public int? OldPrice { get; set; }
+    public decimal? OldPrice { get; set; }
 
     [StringLength(500)]
     public string? About { get; set; }
